image-search/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy<2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2+cpu
git+https://github.com/openai/CLIP.git
pillow